// client/tokens/generated-design-tokens.ts
export const generatedThemes = {
  "dark": {
    "radius-m": 12,
    "radius-full": 100,
    "bottom-bar": "#BEC8C9",
    "bottom-bar-selected": "#33FFFF",
    "container-background": "#002022",
    "container-stroke": "#00363A",
    "container-text": "#BEC8C8",
    "container-icon": "#BEC8C8",
    "background": "#0E1415",
    "primary-button": "#33FFFF",
    "primary-button-text": "#00363A",
    "primary-button-text-pressed": "#DEE3E5",
    "secondary-button": "#002B2E",
    "secondary-button-stroke": "#004347",
    "secondary-button-text": "#80D4DB",
    "brand-name-color": "#FFFFFF",
    "subtitle": "#FA5A4A",
    "text": "#BEC8C9",
    "error message": "#8C0738",
    "device-elements": "#DEE3E5"
  },
  "light": {
    "radius-m": 12,
    "radius-full": 100,
    "bottom-bar": "#002B2E",
    "bottom-bar-selected": "#002B2E",
    "container-background": "#E9F3F3",
    "container-stroke": "#DAE4E4",
    "container-text": "#171D1E",
    "container-icon": "#171D1E",
    "background": "#F5FAFB",
    "primary-button": "#004F54",
    "primary-button-text": "#C5FAFF",
    "primary-button-text-pressed": "#C5FAFF",
    "secondary-button": "#E9F3F3",
    "secondary-button-stroke": "#DAE4E4",
    "secondary-button-text": "#004F54",
    "brand-name-color": "#171D1E",
    "subtitle": "#910809",
    "text": "#3F4849",
    "error message": "#FD5274",
    "device-elements": "#171D1E"
  }
} as const;

export const generatedTypographyTokens = {
  "brandcyan-brand-name-text": {
    "fontFamily": "Nico Moji",
    "fontWeight": "Regular",
    "fontSize": {
      "value": 32,
      "unit": "px"
    },
    "lineHeight": 32,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-display-large": {
    "fontFamily": "Nico Moji",
    "fontWeight": "Regular",
    "fontSize": {
      "value": 57,
      "unit": "px"
    },
    "lineHeight": 64,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-display-medium": {
    "fontFamily": "Nico Moji",
    "fontWeight": "Regular",
    "fontSize": {
      "value": 45,
      "unit": "px"
    },
    "lineHeight": 52,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-display-small": {
    "fontFamily": "Nico Moji",
    "fontWeight": "Regular",
    "fontSize": {
      "value": 36,
      "unit": "px"
    },
    "lineHeight": 44,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-headline-large": {
    "fontFamily": "Nico Moji",
    "fontWeight": "Regular",
    "fontSize": {
      "value": 32,
      "unit": "px"
    },
    "lineHeight": 40,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-headline-medium": {
    "fontFamily": "Oxanium",
    "fontWeight": "Medium",
    "fontSize": {
      "value": 28,
      "unit": "px"
    },
    "lineHeight": 36,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-headline-small": {
    "fontFamily": "Oxanium",
    "fontWeight": "Regular",
    "fontSize": {
      "value": 24,
      "unit": "px"
    },
    "lineHeight": 32,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-body-large": {
    "fontFamily": "Montserrat Alternates",
    "fontWeight": "Medium",
    "fontSize": {
      "value": 16,
      "unit": "px"
    },
    "lineHeight": 24,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-body-medium": {
    "fontFamily": "Montserrat Alternates",
    "fontWeight": "Medium",
    "fontSize": {
      "value": 14,
      "unit": "px"
    },
    "lineHeight": 22,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-body-small": {
    "fontFamily": "Montserrat Alternates",
    "fontWeight": "Medium",
    "fontSize": {
      "value": 12,
      "unit": "px"
    },
    "lineHeight": 18,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-label-large": {
    "fontFamily": "Oxanium",
    "fontWeight": "SemiBold",
    "fontSize": {
      "value": 20,
      "unit": "px"
    },
    "lineHeight": 28,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-label-medium": {
    "fontFamily": "Oxanium",
    "fontWeight": "SemiBold",
    "fontSize": {
      "value": 16,
      "unit": "px"
    },
    "lineHeight": 24,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-label-small": {
    "fontFamily": "Oxanium",
    "fontWeight": "Medium",
    "fontSize": {
      "value": 11,
      "unit": "px"
    },
    "lineHeight": 16,
    "letterSpacing": {
      "value": 0,
      "unit": "px"
    }
  },
  "brandcyan-title-large": {
    "fontFamily": "Oxanium",
    "fontWeight": "SemiBold",
    "fontSize": {
      "value": 22,
      "unit": "px"
    },
    "lineHeight": 28,
    "letterSpacing": {
      "value": 1,
      "unit": "px"
    }
  },
  "brandcyan-title-medium": {
    "fontFamily": "Oxanium",
    "fontWeight": "Medium",
    "fontSize": {
      "value": 16,
      "unit": "px"
    },
    "lineHeight": 24,
    "letterSpacing": {
      "value": 0.8,
      "unit": "px"
    }
  },
  "brandcyan-title-small": {
    "fontFamily": "Oxanium",
    "fontWeight": "Medium",
    "fontSize": {
      "value": 14,
      "unit": "px"
    },
    "lineHeight": 20,
    "letterSpacing": {
      "value": 0.6,
      "unit": "px"
    }
  }
} as const;